Add tracer state tests
