+++ apps/remote_control/test/fixtures/compilation_tracers/lib/compilation_tracers/module_references_by_self.ex
@@ -1,0 +1,13 @@
+defmodule CompilationTracers.ModuleReferencesBySelf1 do
+  __MODULE__
+end
+
+defmodule CompilationTracers.ModuleReferencesBySelf2 do
+  __MODULE__
+end
+
+defmodule CompilationTracers.ModuleReferencesBySelf3 do
+  defmodule Child.Some do
+    __MODULE__
+  end
+end
