+++ apps/remote_control/test/fixtures/compilation_tracers/lib/compilation_tracers/references_imported_by_use.ex
@@ -1,0 +1,13 @@
+defmodule CompilationTracers.ReferencesImportedByUse do
+  use CompilationTracers.ReferencesReferenced
+
+  def uses_fun do
+    referenced_fun()
+  end
+
+  def uses_macro(a) do
+    referenced_macro a do
+      :ok
+    end
+  end
+end
